+++ nest.core.infraestructura.db/General/LicenciaConducirEntityConfig.cs
@@ -38,10 +38,10 @@ namespace nest.core.infraestructura.db.General
         }
     }
 
-    public class LicenciaConducirValueGenerator : ValueGenerator<int>
+    public class LicenciaConducirValueGenerator : ValueGenerator<byte>
     {
         public override bool GeneratesTemporaryValues => false;
-        public override int Next(EntityEntry entry) => (int)GeneradorCorrelativo.GetValue(entry.Context, LicenciaConducirEntityConfig.SCHEMA, LicenciaConducirEntityConfig.TABLE);
-        public override async ValueTask<int> NextAsync(EntityEntry entry, CancellationToken cancellationToken = default) => (int)await GeneradorCorrelativo.GetValueAsync(entry.Context, LicenciaConducirEntityConfig.SCHEMA, LicenciaConducirEntityConfig.TABLE, cancellationToken);
+        public override byte Next(EntityEntry entry) => (byte)GeneradorCorrelativo.GetValue(entry.Context, LicenciaConducirEntityConfig.SCHEMA, LicenciaConducirEntityConfig.TABLE);
+        public override async ValueTask<byte> NextAsync(EntityEntry entry, CancellationToken cancellationToken = default) => (byte)await GeneradorCorrelativo.GetValueAsync(entry.Context, LicenciaConducirEntityConfig.SCHEMA, LicenciaConducirEntityConfig.TABLE, cancellationToken);
     }
 }
